add util to make rendering routes easier
